Move sending messages to extra file
